git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk/tiki@5629 b456876b-0849-0410-b77d-98878d47e9d5

+++ lib/newsletters/nllib.php
@@ -31,10 +31,13 @@ class NlLib extends TikiLib {
 		$now = date("U");
 		$query = "insert into `tiki_sent_newsletters`(`nlId`,`subject`,`data`,`sent`,`users`) values(?,?,?,?,?)";
 		$result = $this->query($query,array((int)$nlId,$subject,$data,(int)$now,$users));
+		$query = "update `tiki_newsletters` set `editions`= `editions`+ 1 where `nlId`=?";
+		$result = $this->query($query,array((int)$nlId));
 	}
 
+	/* get only the email subscribers */
 	function get_subscribers($nlId) {
-		$query = "select email from `tiki_newsletter_subscriptions` where `valid`=? and `nlId`=?";
+		$query = "select `email` from `tiki_newsletter_subscriptions` where `valid`=? and `nlId`=? and isUser !='y'";
 		$result = $this->query($query, array('y',(int)$nlId));
 		$ret = array();
 		while ($res = $result->fetchRow()) {
@@ -43,33 +46,107 @@ class NlLib extends TikiLib {
 		return $ret;
 	}
 
-	function remove_newsletter_subscription($nlId, $email) {
-		$valid = $this->getOne("select `valid` from `tiki_newsletter_subscriptions` where `nlId`=? and `email`=?", array((int)$nlId,$email));
-		$query = "delete from `tiki_newsletter_subscriptions` where `nlId`=? and `email`=?";
-		$result = $this->query($query, array((int)$nlId,$email));
-		$this->update_users($nlId);
+	function get_all_subscribers($nlId, $genUnsub) {
+		global $userlib;
+			/* cleaning */
+		if ($genUnsub == "y") {	
+			$query = "delete from `tiki_newsletter_subscriptions` where `nlId`=? and `isUser`=? and `valid` != ?";
+			$result = $this->query($query, array((int)$nlId, 'g','x'));
+		}
+			/* get the root groups */
+		$query = "select `groupName` from `tiki_newsletter_groups` where `nlId`=?";
+		$result = $this->query($query, array((int)$nlId));
+		$ret = array();
+		$groups =array();
+		while ($res = $result->fetchRow()) {
+			$groups = array_merge($groups, array($res["groupName"]), $this->get_groups_all($res["groupName"]));
+		}
+		 	/* users already in or unsubscribed */
+		$query = "select * from `tiki_newsletter_subscriptions` where `nlId`=? and `isUser`!='g'";
+		$result = $this->query($query, array((int)$nlId));
+		$out = array();
+		$in = array();
+		$inEmail = array();
+		$potential = array();
+		while ($res = $result->fetchRow()) {
+			if ($res["valid"] == "x" || $res["valid"] == "n")
+				$out[] = $res["email"];
+			elseif ($res["isUser"] == "n" && $res["valid"] == "y") {
+				$inEmail[] = $res["email"]; 
+				$ret[] = array("login"=>"", "email"=>$res["email"], "code"=>$res["code"]);
+			} elseif ($res["isUser"] == "y" && $res["valid"] == "y") {
+				$potential[] = $res;
+			}
+		}
+		foreach ($potential as $res) {
+			$in[] = $res["email"];
+			$email = $userlib->get_user_email($res["email"]);
+			if (!in_array($email, $inEmail) && !in_array($email, $out))
+				$ret[] = array("login"=>$res["email"], "email"=>$email, "code"=>$res["code"]);
+		}
+			/* potential users */
+		if (count($groups) > 0) {
+			$mid = " and (".implode(" or ",array_fill(0,count($groups),"`groupName`=?")).")";
+			$query = "select  distinct uu.`login`, uu.`email` from `users_users` uu, `users_usergroups` ug where uu.`userId`=ug.`userId` ".$mid; 
+			$result = $this->query($query, $groups);
+			while ($res = $result->fetchRow()) {
+				if (!in_array($res["login"], $in) && !in_array($res["login"], $out) && !in_array($res["email"], $inEmail)) {
+					if ($genUnsub == "y") {
+						$query = "insert into `tiki_newsletter_subscriptions`(`nlId`,`email`,`code`,`valid`,`subscribed`,`isUser`) values(?,?,?,?,?,?)";
+						$code = $this->genRandomString($res["login"]);
+						$this->query($query,array((int)$nlId,$res["login"],$code,'y',(int)date('U'),'g'));
+						$res["code"] = $code;
+					}
+					$ret[] = $res;
+				}
+			}
+		}
+//echo "<pre>Subscribers:";print_r($ret); echo "</pre>";
+		return $ret;
 	}
 
-	function newsletter_subscribe($nlId, $email, $charset="utf-8") {
-		global $smarty;
-		global $tikilib;
-		global $user;
-		global $sender_email;
-		$info = $this->get_newsletter($nlId);
-		$smarty->assign('info', $info);
-		$code = $this->genRandomString($email);
+	function remove_newsletter_subscription($nlId, $email, $isUser) {
+		$query = "delete from `tiki_newsletter_subscriptions` where `nlId`=? and `email`=? and `isUser`=?";
+		$result = $this->query($query, array((int)$nlId,$email, $isUser),-1, -1, false);
+		/*$this->update_users($nlId);*/
+	}
+
+	function remove_newsletter_group($nlId, $group) {
+		$query = "delete from `tiki_newsletter_groups` where `nlId`=? and `groupName`=?";
+		$result = $this->query($query, array((int)$nlId,$group), -1, -1, false);
+	}
+
+	function newsletter_subscribe($nlId, $add, $isUser='n', $validateAddr='', $addEmail='') {
+		global $smarty, $tikilib, $user, $sender_email,  $userlib;
+		if (empty($add))
+			return false;
+		if ($isUser == "y" && $addEmail == "y") {
+			$add = $userlib->get_user_email($add);
+			$isUser="n";
+		}
+		$query = "select * from `tiki_newsletter_subscriptions` where `nlId`=? and `email`=? and `isUser`=? and `valid`=?";
+		$result = $this->query($query,array((int)$nlId,$add,$isUser, 'n'));
+		if ($res = $result->fetchRow()) {
+			return false; /* already subscribed and valid - keep the same valid status */
+			}
+		$code = $this->genRandomString($add);
 		$now = date("U");
-		if ($info["validateAddr"] == 'y') {
+		$info = $this->get_newsletter($nlId);
+		if ($info["validateAddr"] == 'y' && $validateAddr != 'n') {
+			if ($isUser == "y")
+				$email = $userlib->get_user_email($add);
+			else
+				$email = $add;
+			/* if already has validated don't ask again */
 			// Generate a code and store it and send an email  with the
 			// URL to confirm the subscription put valid as 'n'
 			$foo = parse_url($_SERVER["REQUEST_URI"]);
 			$foopath = preg_replace('/tiki-admin_newsletter_subscriptions.php/', 'tiki-newsletters.php', $foo["path"]);
 			$url_subscribe = $tikilib->httpPrefix(). $foopath;
-			$query = "delete from `tiki_newsletter_subscriptions` where `nlId`=? and `email`=?";
-			$result = $this->query($query,array((int)$nlId,$email));
-			$query = "insert into `tiki_newsletter_subscriptions`(`nlId`,`email`,`code`,`valid`,`subscribed`) values(?,?,?,?,?)";
-			$result = $this->query($query,array((int)$nlId,$email,$code,'n',(int)$now));
+			$query = "insert into `tiki_newsletter_subscriptions`(`nlId`,`email`,`code`,`valid`,`subscribed`,`isUser`) values(?,?,?,?,?,?)";
+			$result = $this->query($query,array((int)$nlId,$add,$code,'n',(int)$now,$isUser));
 			// Now send an email to the address with the confirmation instructions
+			$smarty->assign('info', $info);
 			$smarty->assign('mail_date', date("U"));
 			$smarty->assign('mail_user', $user);
 			$smarty->assign('code', $code);
@@ -84,14 +161,13 @@ class NlLib extends TikiLib {
 			$mail->setText($mail_data);
 			if (!$mail->send(array($email)))
 				return false;
+			return true;
 		} else {
-			$query = "delete from `tiki_newsletter_subscriptions` where `nlId`=? and `email`=?";
-			$result = $this->query($query,array((int)$nlId,$email));
-			$query = "insert into `tiki_newsletter_subscriptions`(`nlId`,`email`,`code`,`valid`,`subscribed`) values(?,?,?,?,?)";
-			$result = $this->query($query,array((int)$nlId,$email,$code,'y',(int)$now));
+			$query = "insert into `tiki_newsletter_subscriptions`(`nlId`,`email`,`code`,`valid`,`subscribed`,`isUser`) values(?,?,?,?,?,?)";
+			$result = $this->query($query,array((int)$nlId,$add,$code,'y',(int)$now,$isUser));
 		}
-		$this->update_users($nlId);
-		return true;
+		/*$this->update_users($nlId);*/
+		return false;
 	}
 
 	function confirm_subscription($code) {
@@ -114,7 +190,13 @@ class NlLib extends TikiLib {
 		$result = $this->query($query,array('y',$code));
 		// Now send a welcome email
 		$smarty->assign('mail_date', date("U"));
-		$user = $userlib->get_user_by_email($res["email"]); //global $user is not necessary defined as the user is not necessary logged in
+		if ($res["isUser"] == "y") {
+			$user = $res["email"];
+			$email = $userlib->get_user_email($user);
+		} else {
+			$email = $res["email"];
+			$user = $userlib->get_user_by_email($email); //global $user is not necessary defined as the user is not necessary logged in
+		}
 		$smarty->assign('mail_user', $user);
 		$smarty->assign('code', $res["code"]);
 		$smarty->assign('url_subscribe', $url_subscribe);
@@ -127,7 +209,7 @@ class NlLib extends TikiLib {
 		$mail->setSubject(sprintf($mail_data, $info["name"], $_SERVER["SERVER_NAME"]));
 		$mail_data = $smarty->fetchLang($lg, 'mail/newsletter_welcome.tpl');
 		$mail->setText($mail_data);
-		if (!$mail->send(array($res["email"])))
+		if (!$mail->send(array($email)))
 				return false;
 		return $this->get_newsletter($res["nlId"]);
 	}
@@ -149,11 +231,20 @@ class NlLib extends TikiLib {
 		$info = $this->get_newsletter($res["nlId"]);
 		$smarty->assign('info', $info);
 		$smarty->assign('code', $res["code"]);
-		$query = "delete from `tiki_newsletter_subscriptions` where `code`=?";
-		$result = $this->query($query,array($code));
+		if ($res["isUser"] == 'g')
+			$query = "update `tiki_newsletter_subscriptions` set `valid`='x' where `code`=?";
+		else
+			$query = "delete from `tiki_newsletter_subscriptions` where `code`=?";
+		$result = $this->query($query,array($code), -1, -1, false);
 		// Now send a bye bye email
 		$smarty->assign('mail_date', date("U"));
-		$user = $userlib->get_user_by_email($res["email"]); //global $user is not necessary defined as the user is not necessary logged in
+		if ($res["isUser"] == "y") {
+			$user = $res["email"];
+			$email = $userlib->get_user_email($user);
+		} else {
+			$email = $res["email"];
+			$user = $userlib->get_user_by_email($email); //global $user is not necessary defined as the user is not necessary logged in
+		}
 		$smarty->assign('mail_user', $user);
 		$smarty->assign('url_subscribe', $url_subscribe);
 		$lg = !$user? $language: $this->get_user_preference($user, "language", $language);
@@ -165,42 +256,56 @@ class NlLib extends TikiLib {
 		$mail->setSubject(sprintf($mail_data, $info["name"], $_SERVER["SERVER_NAME"]));
 		$mail_data = $smarty->fetchLang($lg, 'mail/newsletter_byebye.tpl');
 		$mail->setText($mail_data);
-		$mail->send(array($res["email"]));
+		$mail->send(array($email));
 
-		$this->update_users($res["nlId"]);
+		/*$this->update_users($res["nlId"]);*/
 		return $this->get_newsletter($res["nlId"]);
 	}
 
-	function add_all_users($nlId) {
-		$query = "select `email` from `users_users`";
+	function add_all_users($nlId, $validateAddr='', $addEmail='') {
+		$query = "select `email`, `login`from `users_users`";
 		$result = $this->query($query,array());
 		while ($res = $result->fetchRow()) {
-			$email = $res["email"];
-			if (!empty($email)) {
-				$this->newsletter_subscribe($nlId, $email);
+			if ($addEmail == "y") {
+				$add = $res["email"];
+				$isUser = "n";
+			} else {
+				$add = $res["login"];
+				$isUser = "y";
+			}
+			if (!empty($add)) {
+				$this->newsletter_subscribe($nlId, $add, $isUser, $validateAddr, $addEmail);
 			}
 		}
 	}
 
-	function add_all_group_emails($nlId,$group) {
+	function add_group($nlId, $group) {
+		$query = "delete from `tiki_newsletter_groups` where `nlId`=? and `groupName`=?";
+		$result = $this->query($query,array((int)$nlId,$group), -1, -1, false);
+		$code = $this->genRandomString($group);
+		$query = "insert into `tiki_newsletter_groups`(`nlId`,`groupName`,`code`) values(?,?,?)";
+		$result = $this->query($query,array((int)$nlId,$group,$code));
+	}
+
+	function add_group_users($nlId, $group, $validateAddr='', $addEmail='') {
 		$groups =  array_merge(array($group),$this->get_groups_all($group));
 		$mid = implode(" or ",array_fill(0,count($groups),"`groupName`=?"));
 		$query = "select `login`,`email`  from `users_users` uu, `users_usergroups` ug where uu.`userId`=ug.`userId` and ($mid)";
 		$result = $this->query($query,$groups);
 		$ret = array();
 		while ($res = $result->fetchRow()) {
-			if (!empty($res['email'])) {
-				// $this->newsletter_subscribe($nlId, $res['email']);
+			if ($addEmail == "y")
 				$ret[] = $res['email'];
-			}
+			else
+				$ret[] = $res['login'];
 		}
 		$ret = array_unique($ret);
+print_r($ret);
+		$isUser = $addEmail == "y"?"n": "y";
 		foreach ($ret as $o) {
-			$this->newsletter_subscribe($nlId, $o);
+			$this->newsletter_subscribe($nlId, $o, $isUser, $validateAddr, $addEmail);
 		}
-		return $ret;
 	}
-
 
 	function get_newsletter($nlId) {
 		$query = "select * from `tiki_newsletters` where `nlId`=?";
@@ -219,12 +324,12 @@ class NlLib extends TikiLib {
 	}
 
 	function update_users($nlId) {
-		$users = $this->getOne("select count(*) from `tiki_newsletter_subscriptions` where `nlId`=?",array((int)$nlId));
+		$users = $this->getOne("select count(*) from `tiki_newsletter_subscriptions` where `nlId`=? and `valid`!=?",array((int)$nlId, 'x'));
 		$query = "update `tiki_newsletters` set `users`=? where `nlId`=?";
 		$result = $this->query($query,array($users,(int)$nlId));
 	}
 
-	function list_newsletters($offset, $maxRecords, $sort_mode, $find) {
+	function list_newsletters($offset, $maxRecords, $sort_mode, $find, $update='') {
 		$bindvars = array();
 		if ($find) {
 			$findesc = '%' . $find . '%';
@@ -242,7 +347,10 @@ class NlLib extends TikiLib {
 		$ret = array();
 
 		while ($res = $result->fetchRow()) {
-			$res["confirmed"] = $this->getOne("select count(*) from `tiki_newsletter_subscriptions` where `valid`=? and `nlId`=?",array('y',(int)$res["nlId"]));
+			$ok = count($this->get_all_subscribers($res["nlId"], ""));
+			$notok = $this->getOne("select count(*) from `tiki_newsletter_subscriptions` where `valid`=? and `nlId`=?",array('n',(int)$res["nlId"]));
+			$res["users"] = $ok + $notok;
+			$res["confirmed"] = $ok;
 			$ret[] = $res;
 		}
 		$retval = array();
@@ -287,13 +395,13 @@ class NlLib extends TikiLib {
 		$bindvars = array((int)$nlId);
 		if ($find) {
 			$findesc = '%' . $find . '%';
-			$mid = " where `nlId`=? and `email` like ?";
+			$mid = " where `nlId`=? and `isUser`!='g' and `email` like ?";
 			$bindvars[] = $findesc;
 		} else {
-			$mid = " where `nlId`=? ";
+			$mid = " where `nlId`=? and `isUser`!='g' ";
 		}
 
-		$query = "select * from `tiki_newsletter_subscriptions` $mid order by ".$this->convert_sortmode("$sort_mode");
+		$query = "select * from `tiki_newsletter_subscriptions` $mid order by ".$this->convert_sortmode("$sort_mode").", email asc";
 		$query_cant = "select count(*) from tiki_newsletter_subscriptions $mid";
 		$result = $this->query($query,$bindvars,$maxRecords,$offset);
 		$cant = $this->getOne($query_cant,$bindvars);
@@ -308,35 +416,86 @@ class NlLib extends TikiLib {
 		return $retval;
 	}
 
-	function get_unsub_msg($nlId, $email, $lang) {
+	function list_newsletter_groups($nlId, $offset=-1, $maxRecords=-1, $sort_mode='groupName_asc', $find='') {
+		$bindvars = array((int)$nlId);
+		if ($find) {
+			$findesc = '%' . $find . '%';
+			$mid = " where `nlId`=? and `groupName` like ?";
+			$bindvars[] = $findesc;
+		} else {
+			$mid = " where `nlId`=? ";
+		}
+
+		$query = "select * from `tiki_newsletter_groups` $mid order by ".$this->convert_sortmode("$sort_mode");
+		$query_cant = "select count(*) from tiki_newsletter_groups $mid";
+		$result = $this->query($query,$bindvars,$maxRecords,$offset);
+		$cant = $this->getOne($query_cant,$bindvars);
+		$ret = array();
+
+		while ($res = $result->fetchRow()) {
+			$ret[] = $res;
+		}
+		$retval = array();
+		$retval["data"] = $ret;
+		$retval["cant"] = $cant;
+		return $retval;
+	}
+
+	function get_unsub_msg($nlId, $email, $lang, $code='', $user='') {
 		global $smarty, $language,$userlib,$tikilib;
 		$foo = parse_url($_SERVER["REQUEST_URI"]);
 
 		$foo = str_replace('send_newsletters', 'newsletters', $foo);
 		$url_subscribe = $tikilib->httpPrefix(). $foo["path"];
-		$code = $this->getOne("select `code` from `tiki_newsletter_subscriptions` where `nlId`=? and `email`=?",array((int)$nlId,$email));
+		if ($code == '') {
+			$isUser = $user? "y": "n";
+			$code = $this->getOne("select `code` from `tiki_newsletter_subscriptions` where `nlId`=? and `email`=? and `isUser`=?",array((int)$nlId, $email, $isUser));
+		}
 		$url_unsub = $url_subscribe . '?unsubscribe=' . $code;
-		$user = $userlib->get_user_by_email($email);
-		$lg = !$user? $language: $this->get_user_preference($user, "language", $language);
-		$msg = $smarty->fetchLang($lg, 'mail/newsletter_unsubscribe.tpl');
-		$msg = "<br />\n--\n<br />" . $msg . ":\n<br />:  <a href='$url_unsub'>$url_unsub</a>";
+		$smarty->assign('url_unsub', $url_unsub);
+		if ($user == '')
+			$user = $userlib->get_user_by_email($email);
+		if ($lang == '')
+			$lang = !$user? $language: $this->get_user_preference($user, "language", $language);
+		$msg = $smarty->fetchLang($lang, 'mail/newsletter_unsubscribe.tpl');
 		return $msg;
 	}
 
 	function remove_newsletter($nlId) {
 		$query = "delete from `tiki_newsletters` where `nlId`=?";
-		$result = $this->query($query,array((int)$nlId));
+		$result = $this->query($query,array((int)$nlId), -1, -1, false);
 		$query = "delete from `tiki_newsletter_subscriptions` where `nlId`=?";
-		$result = $this->query($query,array((int)$nlId));
+		$result = $this->query($query,array((int)$nlId), -1, -1, false);
+		$query = "delete from `tiki_newsletter_groups` where `nlId`=?";
+		$result = $this->query($query,array((int)$nlId), -1, -1, false);
 		$this->remove_object('newsletter', $nlId);
 		return true;
 	}
 
-	function remove_edition($editionId) {
+	function remove_edition($nlId, $editionId) {
 		$query = "delete from `tiki_sent_newsletters` where `editionId`=?";
-		$result = $this->query($query,array((int)$editionId));
+		$result = $this->query($query,array((int)$editionId),-1, -1, false);
+		$query = "update `tiki_newsletters` set `editions`= `editions`- 1 where `nlId`=?";
+		$result = $this->query($query,array((int)$nlId));
 	}
 
+	function valid_subscription($nlId, $email, $isUser) {
+		$query = "update `tiki_newsletter_subscriptions` set `valid`= ? where `nlId`=? and `email`=? and `isUser`=?";
+		$result = $this->query($query, array('y', (int)$nlId, $email, $isUser));
+	}
+
+	function list_tpls() {
+		global $tikidomain;
+		$tpls = array();
+		if (is_dir("templates/newsletters/$tikidomain")) {
+			$h = opendir("templates/newsletters/$tikidomain");
+ 			while ($file = readdir($h)) {
+				if (ereg("\.tpl$", $file))
+					$tpls[] = $file;
+			}
+		}
+		return $tpls;
+	}
 }
 
 $nllib = new NlLib($dbTiki);
